Added bank + fee estimation traits
