+++ cw-orch-daemon/src/sync/mod.rs
@@ -1,4 +1,4 @@
 mod builder;
 mod core;
-
+mod modules;
 pub use self::{builder::*, core::*};
