+++ cw-orch/src/osmosis_test_tube/mod.rs
@@ -2,5 +2,5 @@
 //! It has an associated state that stores deployment information for easy retrieval and contract interactions.
 
 mod core;
-
+mod modules;
 pub use self::core::*;
