+++ packages/cw-orch-core/src/environment/cosmwasm_environment.rs
@@ -1,9 +1,8 @@
 //! Transactional traits for execution environments.
 
-use super::{modules::Bank, ChainState, IndexResponse};
+use super::{ChainState, IndexResponse};
 use crate::{contract::interface_traits::Uploadable, error::CwEnvError};
 use cosmwasm_std::{Addr, BlockInfo, Coin};
-use dialoguer::Confirm;
 use serde::{de::DeserializeOwned, Serialize};
 use std::fmt::Debug;
 
@@ -80,61 +79,4 @@ pub trait TxHandler: ChainState + Clone {
         new_code_id: u64,
         contract_address: &Addr,
     ) -> Result<Self::Response, Self::Error>;
-}
-
-pub struct WalletBalanceAssertionResult {
-    pub assertion: bool,
-    pub expected: Coin,
-    pub current: Coin,
-}
-
-pub trait WalletBalanceAssertion: TxHandler {
-    /// An internal function to check wether the wallet has enough balance to execute an action
-    fn _assert_wallet_balance(
-        &self,
-        gas: u64,
-    ) -> Result<WalletBalanceAssertionResult, <Self as TxHandler>::Error>;
-
-    /// The result of a bad wallet balance on the program execution
-    /// You don't need to reimplement this method
-    fn assert_wallet_balance(&self, gas: u64) -> Result<(), CwEnvError> {
-        let result = self._assert_wallet_balance(gas).map_err(Into::into)?;
-
-        if result.assertion {
-            log::debug!("The wallet has enough balance to deploy");
-            return Ok(());
-        }
-
-        // Needs to be pushed on the daemon impl
-        // log::debug!(
-        //     "Checking balance {} on chain {}, address {}. Expecting {}{}",
-        //     balance.amount,
-        //     chain_info.chain_id,
-        //     chain.sender(),
-        //     fee,
-        //     fee_token.denom.as_str()
-        // );
-
-        // If there is not enough asset balance, we need to warn the user
-        if Confirm::new()
-            .with_prompt(format!(
-                "Not enough funds on chain {} at address {} to deploy the contract. 
-                    Needed: {} but only have: {}.
-                    Press 'y' when the wallet balance has been increased to resume deployment",
-                self.block_info().map_err(Into::into)?.chain_id,
-                self.sender(),
-                result.expected,
-                result.current
-            ))
-            .interact()?
-        {
-            // We retry asserting the balance
-            self.assert_wallet_balance(gas)
-        } else {
-            Err(CwEnvError::NotEnoughBalance {
-                expected: result.expected,
-                current: result.current,
-            })
-        }
-    }
 }
