+++ packages/cw-orch-core/src/environment/modules/bank.rs
@@ -1,37 +1,13 @@
-use cosmwasm_std::Coin;
+use cosmwasm_std::{Coin, Addr};
 
-use super::fee_estimation::FeeEstimation;
+use crate::environment::TxHandler;
 
 /// Allows to send bank related transactions and query the on-chain bank module
-pub trait Bank<Response, Error>: FeeEstimation<Error> + std::fmt::Debug {
+pub trait Bank: TxHandler {
     // Send coins to another account
-    fn send(&self, funds: Vec<Coin>, receiver: Self) -> Result<Response, Error>;
+    fn send(&self, recipient: Addr, funds: Vec<Coin>) -> Result<<Self as TxHandler>::Response, <Self as TxHandler>::Error>;
 
     // Send coins to another account
-    fn balance(&self, denom: Option<String>) -> Result<Vec<Coin>, Error>;
+    fn balance(&self, denom: Option<String>) -> Result<Vec<Coin>, <Self as TxHandler>::Error>;
 
-    fn wallet_balance_assertion(&self, gas: u64) -> Result<WalletBalanceAssertionResult, Error> {
-        let estimated_fee = self.estimate_fee(gas)?;
-
-        let balance = self.balance(Some(estimated_fee.denom.clone()))?;
-
-        log::debug!(
-            "Checking balance {:?}, address {:?}. Expecting {:?}",
-            balance.clone(),
-            self,
-            estimated_fee.clone()
-        );
-
-        Ok(WalletBalanceAssertionResult {
-            expected: estimated_fee.clone(),
-            current: balance[0].clone(),
-            assertion: balance[0].amount >= estimated_fee.amount,
-        })
-    }
-}
-
-pub struct WalletBalanceAssertionResult {
-    pub assertion: bool,
-    pub expected: Coin,
-    pub current: Coin,
 }
