+++ packages/cw-orch-core/src/environment/modules/mod.rs
@@ -1,3 +1,3 @@
 mod bank;
-mod fee_estimation;
+pub mod fee_estimation;
 pub use bank::Bank;
